fix: run startup handlers after all variables have been processed
If the startup handlers are run before them, then even though the code
does initialise variables, it does it in the wrong order. For example,
for `v = initMembPot` that is a `StateAssignment` that should happen
`OnStart`, `initMembPot` needs to be processed first so that `v` can
then pick up its value.

Fixes https://github.com/LEMS/pylems/issues/20